[Refactor] #267 - 유저 회원가입 시 발행되는 이벤트 비동기 로직으로 변경
