+++ moonshot-api/src/main/java/org/moonshot/user/service/UserSignUpService.java
@@ -1,0 +1,31 @@
+package org.moonshot.user.service;
+
+import java.time.LocalDateTime;
+import lombok.RequiredArgsConstructor;
+import org.moonshot.discord.SignUpEvent;
+import org.moonshot.user.model.User;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class UserSignUpService {
+
+    private final ApplicationEventPublisher eventPublisher;
+
+    @Async
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    public void publishSignUpEvent(User user) {
+        eventPublisher.publishEvent(SignUpEvent.of(
+                user.getName(),
+                user.getEmail() == null ? "" : user.getEmail(),
+                user.getSocialPlatform().toString(),
+                LocalDateTime.now(),
+                user.getImageUrl()
+        ));
+    }
+
+}
